+++ src/Subscriber/CategoryPageSubscriber.php
@@ -12,7 +12,7 @@ use Shopware\Storefront\Page\Navigation\NavigationPageLoadedEvent;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 use function Omikron\FactFinder\Shopware6\Internal\Utils\safeGetByName;
 
-class CategoryView implements EventSubscriberInterface
+class CategoryPageSubscriber implements EventSubscriberInterface
 {
     private AbstractCategoryRoute $cmsPageRoute;
 
@@ -26,12 +26,12 @@ class CategoryView implements EventSubscriberInterface
         AbstractCategoryRoute $cmsPageRoute,
         CategoryPath $categoryPath,
         string $categoryPathFieldName,
-        array $initialNavigationParams = []
+        array $categoryPageAddParams = []
     ) {
         $this->cmsPageRoute = $cmsPageRoute;
         $this->categoryPath = $categoryPath;
         $this->fieldName    = $categoryPathFieldName;
-        $this->addParams    = $initialNavigationParams;
+        $this->addParams    = $categoryPageAddParams;
     }
 
     public static function getSubscribedEvents()
@@ -41,8 +41,8 @@ class CategoryView implements EventSubscriberInterface
 
     public function onPageLoaded(NavigationPageLoadedEvent $event): void
     {
-        $navigationId     = $event->getRequest()->get('navigationId', $event->getSalesChannelContext()->getSalesChannel()->getNavigationCategoryId());
-        $category         = $this->cmsPageRoute->load($navigationId, $event->getRequest(), $event->getSalesChannelContext())->getCategory();
+        $navigationId = $event->getRequest()->get('navigationId', $event->getSalesChannelContext()->getSalesChannel()->getNavigationCategoryId());
+        $category     = $this->cmsPageRoute->load($navigationId, $event->getRequest(), $event->getSalesChannelContext())->getCategory();
 
         $disableImmediate = safeGetByName($category->getCustomFields())(OmikronFactFinder::DISABLE_SEARCH_IMMEDIATE_CUSTOM_FIELD_NAME);
         $isHome           = $event->getRequest()->get('_route') === 'frontend.home.page';
@@ -50,9 +50,19 @@ class CategoryView implements EventSubscriberInterface
 
         $categoryPath     = $this->getPath($this->cmsPageRoute->load($navigationId, $event->getRequest(), $event->getSalesChannelContext())->getCategory());
 
+        $baseAddParams   = array_filter(explode(',', (string) safeGetByName($event->getPage()->getExtension('factfinder')->get('communication'))('add-params')));
+        /**
+         * $this->addParams + $baseAddParams will not override entries from $baseAddParams as $this->addParams is associative array and $baseAddParams is not
+         * $this->addParams is associative array because this is how parameters collection is passed as an argument constructor in Symfony.
+         */
+        $mergedAddParams = array_reduce($this->addParams + $baseAddParams, function (array $acc, string $expr): array {
+            list($key, $value) = explode('=', $expr);
+            return $acc + [$key => $value];
+        }, []);
+
         $communication = [
                 'search-immediate' => !$isHome && !$disableImmediate ? 'true' : 'false',
-                'add-params'       => implode(',', $this->addParams),
+                'add-params'       => implode(',', array_map(fn (string $key, string $value) => sprintf('%s=%s', $key, $value), array_keys($mergedAddParams), array_values($mergedAddParams))),
             ] + ($isCategory ? ['category-page' => $this->prepareCategoryPath($categoryPath)] : []);
 
         $event->getPage()->getExtension('factfinder')->assign(['communication' => $communication]);