+++ src/Subscriber/ConfigurationSubscriber.php
@@ -17,11 +17,18 @@ class ConfigurationSubscriber implements EventSubscriberInterface
 
     private array $communicationParameters;
 
-    public function __construct(Communication $config, array $fieldRoles, array $communicationParameters)
-    {
+    private array $addParams;
+
+    public function __construct(
+        Communication $config,
+        array $fieldRoles,
+        array $communicationParameters,
+        array $configurationAddParams = []
+    ) {
         $this->config                  = $config;
         $this->fieldRoles              = $fieldRoles;
         $this->communicationParameters = $communicationParameters;
+        $this->addParams               = $configurationAddParams;
     }
 
     public static function getSubscribedEvents()
@@ -33,19 +40,24 @@ class ConfigurationSubscriber implements EventSubscriberInterface
     {
         $customer       = $event->getSalesChannelContext()->getCustomer();
         $salesChannelId = $event->getSalesChannelContext()->getSalesChannel()->getId();
+        $communication  = [
+            'url'                   => $this->config->getServerUrl(),
+            'channel'               => $this->config->getChannel($salesChannelId),
+            'version'               => 'ng',
+            'api'                   => 'v4',
+            'user-id'               => $customer ? $customer->getId() : null,
+            'currency-code'         => $event->getSalesChannelContext()->getCurrency()->getIsoCode(),
+            'currency-country-code' => $event->getRequest()->getLocale(),
+            'search-immediate'      => strpos($event->getRequest()->get('_route'), 'factfinder') ? 'true' : 'false',
+        ];
+
+        if (!empty($this->addParams)) {
+            $communication['add-params'] = implode(',', $this->addParams);
+        }
 
         $event->getPage()->addExtension('factfinder', new ArrayEntity([
             'field_roles'   => $this->fieldRoles,
-            'communication' => [
-                'url'                   => $this->config->getServerUrl(),
-                'channel'               => $this->config->getChannel($salesChannelId),
-                'version'               => 'ng',
-                'api'                   => 'v4',
-                'user-id'               => $customer ? $customer->getId() : null,
-                'currency-code'         => $event->getSalesChannelContext()->getCurrency()->getIsoCode(),
-                'currency-country-code' => $event->getRequest()->getLocale(),
-                'search-immediate'      => strpos($event->getRequest()->get('_route'), 'factfinder') ? 'true' : 'false',
-            ] + $this->communicationParameters,
+            'communication' => $communication + $this->communicationParameters,
         ]));
     }
 }